Breaks After Topic Added
it works until a topic is added
rating need to be reformatted
inspector shows class that should be getting click events to load new gifs is being applied to new topic buttons,
but it's not pulling the gifs w/ the topic name.
wanted to save this checkpoint just in case things get even worse.